Separe Post's Categories from Post
New Controller
New Manager
New Entitie
Update the other AND router
